feat:モンスターの画像追加
