+++ src/routes/Routers.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import { Title } from '../pages/Title';
 import { MainMenu } from '../pages/MainMenu';
-import { BeginnerDungeon } from '../pages/BeginnerDungeon';
+import { BeginnerDungeon01 } from '../pages/BeginnerDungeon/BeginnerDungeon01';
 
 export const Routers = () => {
   return (
@@ -9,7 +9,7 @@ export const Routers = () => {
       <Routes>
         <Route path={'/'} element={<Title />} />
         <Route path={'/MainMenu'} element={<MainMenu />} />
-        <Route path={'BeginnerDungeon'} element={<BeginnerDungeon />} />
+        <Route path={'BeginnerDungeon01'} element={<BeginnerDungeon01 />} />
       </Routes>
     </BrowserRouter>
   );
